requirements.txt
pyyaml
watchdog
pywin32; sys_platform == "win32"